feat: Refactor segmentation utilities and enhance worker data handling
- Introduce utilsForWorker module to centralize worker-related functions.
- Refactor getSegmentLargestBidirectional and getStatistics to utilize new utility functions for improved data handling.
- Streamline segmentation data retrieval and processing for worker tasks.
- Update progress triggering to use worker type constants for clarity.
- Remove redundant code and improve overall maintainability.
