+++ packages/tools/src/utilities/segmentation/getStatistics.ts
@@ -1,38 +1,19 @@
+import { utilities, getWebWorkerManager } from '@cornerstonejs/core';
 import {
-  cache,
-  utilities,
-  getWebWorkerManager,
-  eventTarget,
-  Enums,
-  triggerEvent,
-  metaData,
-} from '@cornerstonejs/core';
-import { getActiveSegmentIndex } from '../../stateManagement/segmentation/getActiveSegmentIndex';
-import VolumetricCalculator from './VolumetricCalculator';
-import { getStrategyData } from '../../tools/segmentation/strategies/utils/getStrategyData';
+  triggerWorkerProgress,
+  getSegmentationDataForWorker,
+  prepareVolumeStrategyDataForWorker,
+  prepareStackDataForWorker,
+  getImageReferenceInfo,
+} from './utilsForWorker';
 import { getPixelValueUnitsImageId } from '../getPixelValueUnits';
-import ensureSegmentationVolume from '../../tools/segmentation/strategies/compositions/ensureSegmentationVolume';
-import ensureImageVolume from '../../tools/segmentation/strategies/compositions/ensureImageVolume';
-import { getSegmentation } from '../../stateManagement/segmentation/getSegmentation';
-import { registerComputeWorker } from '../registerComputeWorker';
+import VolumetricCalculator from './VolumetricCalculator';
 import { WorkerTypes } from '../../enums';
-import type {
-  LabelmapSegmentationDataStack,
-  LabelmapSegmentationDataVolume,
-} from '../../types/LabelmapTypes';
 import type { NamedStatistics } from '../../types';
+import { registerComputeWorker } from '../registerComputeWorker';
 
 // Radius for a volume of 10, eg 1 cm^3 = 1000 mm^3
 const radiusForVol1 = Math.pow((3 * 1000) / (4 * Math.PI), 1 / 3);
-
-const workerManager = getWebWorkerManager();
-
-const triggerWorkerProgress = (eventTarget, progress) => {
-  triggerEvent(eventTarget, Enums.Events.WEB_WORKER_PROGRESS, {
-    progress,
-    type: WorkerTypes.COMPUTE_STATISTICS,
-  });
-};
 
 /**
  * Get statistics for a segmentation.
@@ -61,44 +42,21 @@ async function getStatistics({
 }): Promise<NamedStatistics | { [segmentIndex: number]: NamedStatistics }> {
   registerComputeWorker();
 
-  triggerWorkerProgress(eventTarget, 0);
+  triggerWorkerProgress(WorkerTypes.COMPUTE_STATISTICS, 0);
 
-  const segmentation = getSegmentation(segmentationId);
-  const { representationData } = segmentation;
+  const segData = getSegmentationDataForWorker(segmentationId, segmentIndices);
 
-  const { Labelmap } = representationData;
-
-  if (!Labelmap) {
-    console.debug('No labelmap found for segmentation', segmentationId);
+  if (!segData) {
     return;
   }
 
-  const segVolumeId = (Labelmap as LabelmapSegmentationDataVolume).volumeId;
-  const segImageIds = (Labelmap as LabelmapSegmentationDataStack).imageIds;
-
-  // Create a minimal operationData object
-  const operationData = {
-    segmentationId,
-    volumeId: segVolumeId,
-    imageIds: segImageIds,
-  };
-
-  let reconstructableVolume = false;
-  if (segImageIds) {
-    const refImageIds = segImageIds.map((imageId) => {
-      const image = cache.getImage(imageId);
-      return image.referencedImageId;
-    });
-    reconstructableVolume = utilities.isValidVolume(refImageIds);
-  }
-  let indices = segmentIndices;
-
-  if (!indices) {
-    indices = [getActiveSegmentIndex(segmentationId)];
-  } else if (!Array.isArray(indices)) {
-    // Include the preview index
-    indices = [indices, 255];
-  }
+  const {
+    operationData,
+    segVolumeId,
+    segImageIds,
+    reconstructableVolume,
+    indices,
+  } = segData;
 
   // Get reference image ID and modality unit options
   const { refImageId, modalityUnitOptions } = getImageReferenceInfo(
@@ -134,15 +92,7 @@ async function calculateVolumeStatistics({
   mode,
 }) {
   // Get the strategy data
-  const strategyData = getStrategyData({
-    operationData,
-    strategy: {
-      ensureSegmentationVolumeFor3DManipulation:
-        ensureSegmentationVolume.ensureSegmentationVolumeFor3DManipulation,
-      ensureImageVolumeFor3DManipulation:
-        ensureImageVolume.ensureImageVolumeFor3DManipulation,
-    },
-  });
+  const strategyData = prepareVolumeStrategyDataForWorker(operationData);
 
   const {
     segmentationVoxelManager,
@@ -161,8 +111,6 @@ async function calculateVolumeStatistics({
   const segmentationScalarData =
     segmentationVoxelManager.getCompleteScalarDataArray();
 
-  const imageScalarData = imageVoxelManager.getCompleteScalarDataArray();
-
   const segmentationInfo = {
     scalarData: segmentationScalarData,
     dimensions: segmentationImageData.getDimensions(),
@@ -172,14 +120,14 @@ async function calculateVolumeStatistics({
   };
 
   const imageInfo = {
-    scalarData: imageScalarData,
+    scalarData: imageVoxelManager.getCompleteScalarDataArray(),
     dimensions: imageData.getDimensions(),
     spacing: imageData.getSpacing(),
     origin: imageData.getOrigin(),
     direction: imageData.getDirection(),
   };
 
-  const stats = await workerManager.executeTask(
+  const stats = await getWebWorkerManager().executeTask(
     'compute',
     'calculateSegmentsStatisticsVolume',
     {
@@ -190,7 +138,7 @@ async function calculateVolumeStatistics({
     }
   );
 
-  triggerWorkerProgress(eventTarget, 100);
+  triggerWorkerProgress(WorkerTypes.COMPUTE_STATISTICS, 100);
 
   if (mode === 'collective') {
     return processSegmentationStatistics({
@@ -298,36 +246,13 @@ const processSegmentationStatistics = ({
  * Calculate statistics for a stack of images
  */
 async function calculateStackStatistics({ segImageIds, indices, unit, mode }) {
-  triggerWorkerProgress(eventTarget, 0);
-  // we need to loop over each seg image separately and calculate the stats
-  const segmentationInfo = [];
-  const imageInfo = [];
-  for (const segImageId of segImageIds) {
-    const segImage = cache.getImage(segImageId);
-    const segPixelData = segImage.getPixelData();
-    const segVoxelManager = segImage.voxelManager;
-    const segSpacing = [segImage.rowPixelSpacing, segImage.columnPixelSpacing];
+  triggerWorkerProgress(WorkerTypes.COMPUTE_STATISTICS, 0);
 
-    const refImageId = segImage.referencedImageId;
-    const refImage = cache.getImage(refImageId);
-    const refPixelData = refImage.getPixelData();
-    const refVoxelManager = refImage.voxelManager;
-    const refSpacing = [refImage.rowPixelSpacing, refImage.columnPixelSpacing];
+  // Get segmentation and image info for each image in the stack
+  const { segmentationInfo, imageInfo } =
+    prepareStackDataForWorker(segImageIds);
 
-    segmentationInfo.push({
-      scalarData: segPixelData,
-      dimensions: segVoxelManager.dimensions,
-      spacing: segSpacing,
-    });
-
-    imageInfo.push({
-      scalarData: refPixelData,
-      dimensions: refVoxelManager.dimensions,
-      spacing: refSpacing,
-    });
-  }
-
-  const stats = await workerManager.executeTask(
+  const stats = await getWebWorkerManager().executeTask(
     'compute',
     'calculateSegmentsStatisticsStack',
     {
@@ -338,7 +263,7 @@ async function calculateStackStatistics({ segImageIds, indices, unit, mode }) {
     }
   );
 
-  triggerWorkerProgress(eventTarget, 100);
+  triggerWorkerProgress(WorkerTypes.COMPUTE_STATISTICS, 100);
 
   const spacing = segmentationInfo[0].spacing;
   const segmentationImageData = segmentationInfo[0];
@@ -405,40 +330,6 @@ function getSphereStats(testMax, radiusIJK, segData, imageVoxels, spacing) {
   });
 
   return VolumetricCalculator.getStatistics({ spacing });
-}
-
-/**
- * Gets the reference image ID and modality unit options based on segmentation data
- * @param segVolumeId - The segmentation volume ID
- * @param segImageIds - The segmentation image IDs
- * @returns Object containing reference image ID and modality unit options
- */
-function getImageReferenceInfo(segVolumeId, segImageIds) {
-  let refImageId;
-
-  if (segVolumeId) {
-    const segmentationVolume = cache.getVolume(segVolumeId);
-    const imageIds = segmentationVolume.imageIds;
-
-    const cachedImage = cache.getImage(imageIds[0]);
-
-    if (cachedImage) {
-      refImageId = cachedImage.referencedImageId;
-    }
-  } else if (segImageIds?.length) {
-    const segImage = cache.getImage(segImageIds[0]);
-    refImageId = segImage.referencedImageId;
-  }
-
-  const refImage = cache.getImage(refImageId);
-  const scalingModule = metaData.get('scalingModule', refImageId);
-
-  const modalityUnitOptions = {
-    isPreScaled: Boolean(refImage.preScale?.scaled),
-    isSuvScaled: typeof scalingModule?.suvbw === 'number',
-  };
-
-  return { refImageId, modalityUnitOptions };
 }
 
 export default getStatistics;
